+++ src/floating-select/floating-select.component.jsx
@@ -9,6 +9,9 @@ export default class FloatingSelect extends FloatingSelectBase {
       {...this.props}
       components={this.state.components}
       menuPlacement="auto"
+      ref={(el) => {
+        this.selectRef = el;
+      }}
       onMenuOpen={this.handleOpenMenu}
       onMenuClose={this.handleCloseMenu}
       styles={styles(this.props.controlHeight)}